modified the tag structure to have many-many tag-organization model, and one-many tagtype-tag model
