+++ app/models/admin_user.py
@@ -11,35 +11,6 @@ class Permission:
     GENERAL = 0x01
     ADMINISTER = 0xff
 
-class Organization(db.Model):
-    __tablename__ = 'organizations'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(64), unique=True)
-    email = db.Column(db.String(64), nullable=False)
-    phone = db.Column(db.Integer, nullable=False)
-    address = db.Column(db.String(500))
-    website_link = db.Column(db.String(120))
-    hours = db.Column(db.String(64))
-    description = db.Column(db.Text)
-    tags = db.relationship("TagAssociation", back_populates="organizations")
-
-class Tag(db.Model):
-    __tablename__ = 'tag'
-    id = db.Column(db.Integer, primary_key=True)
-    tag_name = db.Column(db.String(64), nullable=False)
-    services_offered =  db.Column(db.String(500))
-    accessibility = db.Column(db.String(64))
-    ages = db.Column(db.String(64))
-    organizations = db.relationship("TagAssociation", back_populates="tag")
-
-class TagAssociation(db.Model):
-    __tablename__ = 'tagassociation'
-    id = db.Column(db.Integer, primary_key=True)
-    organiztion_id = db.Column(db.Integer, db.ForeignKey('organizations.id', ondelete='CASCADE'))
-    tag_id = db.Column(db.Integer, db.ForeignKey('tag.id', ondelete='CASCADE'))
-    tag = db.relationship("Tag", back_populates="organizations")
-    organization = db.relationship("Organization", back_populates="tags")
-
 class Role(db.Model):
     id = db.Column(db.Int
     __tablename__ = 'roles'eger, primary_key=True)
